additional changes.

+++ devlopment/WarDrive/Scripts/build.sh
@@ -1,16 +1,24 @@
 #! /bin/sh
 
-echo "$0:begining.."
+echo "usage: ./build.sh [-skip-cmake]"
 
-rm -rf ../Source/TestApp/Build
-mkdir ../Source/TestApp/Build
+echo "$0,$1:begining.."
+
+option1=$1
+if test "$option1" != "-skip-cmake" 
+then
+  rm -rf ../Source/TestApp/Build
+  mkdir ../Source/TestApp/Build
+  echo "running-cmake.."
+  cd ../Source/TestApp/Build
+  cmake ..
+  cd -
+fi 
+
 
 cd ../Source/TestApp/Build
-
 echo "starting TestApp build.."
 
-echo "running-cmake.."
-cmake ..
 
 echo "running-make.."
 make clean
